feat: product edit page
add admin only page, ui

+++ src/service/wishService.cpp
@@ -7,7 +7,7 @@
 Products WishService::getWishs(int userId){
     Products wishs;
     std::optional<Product> item;
-    if (userId > 1){
+    if (userId > 0 && isRepoValid()){
         for(int id: wishRepo->findByUser(userId)){
             item = productRepo->findById(id);
             if (item != std::nullopt){
@@ -21,12 +21,14 @@ Products WishService::getWishs(int userId){
 
 bool WishService::add(int userId, int productId) {
     if (!isRepoValid()){return false;}
+    if (userId <= 0) {return false;}
     if (productRepo->findById(productId) == std::nullopt){return false;}
     return wishRepo->insert(userId, productId);
 }
 
 bool WishService::remove(int userId, int productId) {
     if (!isRepoValid()){return false;}
+    if (userId <= 0) {return false;}
     return wishRepo->remove(userId, productId);
 }
